+++ app_fdroid/src/main/java/org/gkisalatiga/fdroid/lib/AppColors.kt
@@ -18,7 +18,7 @@ class AppColors {
         const val AGENDA_ITEM_TIME_BACKGROUND = 0xff482505
         const val AGENDA_ITEM_CHIP_SELECTED_BACKGROUND = 0xff482505
         const val LIGHT_THEME_WHITE = 0xffffffff
-        const val MAIN_TOP_BAR_COLOR = 0xff482505
+        const val MAIN_TOP_BAR_COLOR = 0xff715446
         const val MAIN_TOP_BAR_CONTENT_COLOR = 0xffffffff
         const val YKB_ARCHIVE_BUTTON_COLOR = 0xff482505